Update README.md
This PR changes the instructions for running Prometheus and Grafana in Docker to use the `--add-host` parameter instead of the `--link` parameter which is deprecated and which has been reported by @shaunsmith to sometimes not work. 

The supplied command has been verified by @shaunsmith on Mac and Linux and by me on Linux.