status: fail to debug system tests
the above patches lay out the path for an event-based invoke path, which
expects a fully qualified cloud event with all protocol/app/fn/trigger
information filled in (to be done at higher levels, in http handler, etc).
at present, no handler is hooked up for it, this is in the midst of plumbing.
the thinking is we turn the call object into a 'convenience' object for the
agent with any configuration extracted from the event that's needed to run the
function, and carry around the event object to re-encode against the
container. did not get to the part where we plumb out the response cloud
event.

with that, there is a requirement to align the agent call path by getting rid
of the http request stuff being baked in so deeply. we were optimizing around
this behavior in a number of paths (lb_agent, protocol dispatch, ...), but if
we have an event only invoke/trigger path (that is, an agent only runs cloud
events) we can't do a lot of those optimizations, at least as easily, since
the data section will be in a json (we can make custom json marshaling stuff,
but we're not really there). this may mean reconsidering some of the lb_agent
optimizations around body (we're gonna have a json blob we have to read out
anyway at the lb level without hackery).

the other thing this does is move the config option for max request size up to
the front end, adds the env var for it there, adds a server test for it and
removes it from agent. a request is either gonna come through the lb (before
grpc) or to the server, we can handle limiting the request there at least now,
which may be easier than having multiple layers of request body checking. this
aligns with not making the agent as responsible for http behaviors
(eventually, not at all once route is fully deprecated).

also took the liberty to take care of some lint nits (AgentConfig)

futzed with the system tests for a couple of hours and I can't quite get to
the bottom of what it doesn't like about this, am mostly learning how it all
works/quirks. the context deadline runs out, so maybe plumbing the ctx out of
get call / submit is causing issues, but I haven't managed to find the source
of the error just yet.

anyway, take it or leave it. it's likely some of this could be exfiltrated
without as much damage (max request body, and first commit just w/ the
handler but not hooked up).