less racy payload thingy
this has all the fine properties of the current (master) version of the
payload cache reader, in that attempting to read the reader into a buffer may
never actually finish, but this time we don't leak our own goroutine
explicitly to do it and just hand out an io.ReadCloser to allow that to
happen. this just allows one thread to start the read-into-buffer process,
after which all threads including that one that are interested in the body are
required to wait. after being read in, the cached body can be handed out
relatively quickly.

we could easily get rid of the closure, we're dragging an extra stack around.
i'm keen on killing this, if you read the comments, the contract here is not
great, you can't actually read the io.reader whenever you want to, even if you
have it, because part of the protocol is actually kicking off the byte stream
from which to read. existence is not worth it some days. alas, this has no
more issues than the previous, while maintaining the GetBody dance we were
doing, without plumbing the request all the way down through the agent.

i'm open to ideas. i'm slightly mortified by this having to be this way, but
it seems slightly better than the last thing in any event.