swap out for github.com/fnproject/cloudevent
this gets rid of the fn core specific event that was added in the previous
commit. we may need to modify the cloudevent lib to our needs, but anyway,
this switches over to it, we can go from here.

added various TODO in the code from things that stuck out that we need to do.
in addition to those:

TODO these will interface with our own hokey protocol/cloudevent thing OK but
we should probably swap that out, too, to be sure.
TODO we need to figure out if we want to fill in some of the 'required' fields
or require users to supply them. with the cli it's pretty easy, and it would
be nice to validate those fields for sanity.