attempt a cohesive cloud events plan
there are very few TODOs, but some. mostly, there are concrete suggestions for
them. to date, this plan exists in old docs in closed down issues that are now
stale, so this is a new plan with a smaller delta that we can follow, with an
attempt to approximate the actual work involved at the bottom.

this is trying to have the simplest possible cloud event format, without us
modifying the things into oblivion (only http trigger does this as proposed)

it seems kinda reasonable if we are willing to commit, so maybe we should
mostly talk about the commitment part, technically speaking there's not a lot
to these things. i have little success in extracating these kinds of things
into an english document, so we'll see how this goes...